refactor(packaging): move effect and form144 into their subpackages, shim the old paths
Second slice of edgartools-07lk.12.1.

  edgar/effect.py   -> edgar/offerings/effect.py   (registration lifecycle:
                       joins S-1/S-3/S-4, 424B, Form C/D)
  edgar/form144.py  -> edgar/ownership/form144.py  (an affiliate's notice of a
                       proposed sale: joins Forms 3/4/5)

WHY A SHIM IS MANDATORY. Neither name is in edgar.__all__ and neither had a
top-level re-export, so `from edgar.form144 import Form144` was the ENTIRE
public surface for these. Moving without a shim is a break, not a staged
deprecation — 07lk.23 requires the additive half to ship in 5.x first.

The shims forward via PEP 562 __getattr__ rather than a re-export list, which
preserves object IDENTITY: edgar.form144.Form144 IS
edgar.ownership.form144.Form144. A shim re-exporting copies would satisfy a
naive import test and then fail every isinstance check written against the
other path. It also cannot drift when a name is added to the real module.

obj()'s lazy dispatch is repointed at the new paths. Left alone it would route
every 144/EFFECT filing through the shim and warn at users doing nothing wrong,
and a deprecation that fires on correct usage is noise people filter out
wholesale.

The four existing test files still import the old paths, deliberately: they now
exercise the shim, which is the part most likely to be silently wrong.

TWO CORRECTIONS TO THE BEAD, both from measuring rather than trusting:

shelfofferings.py is NOT a single-form parser. It is a 12-line
list_takedown_forms() helper belonging beside edgar/forms.py, and its only
consumer patches its internals (patch.object(shelfofferings, "list_forms")),
which a shim would leave aimed at the wrong module object — silently gutting
the test. Left where it is.

muniadvisors needs NO shim: a module converting to a package of the same name
keeps edgar.muniadvisors resolving exactly as before. That move lands next, on
top of #1199's changes to the file.
