+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **Four dead modules that no longer had callers.** `edgar/tools/` (an empty package), `edgar/analysis/` and `edgar/xbrl/analysis/` (both already emptied of their modules, leaving only stale caches), and `edgar/reference/financials.py` (a seven-line scratch script with a `__main__` block and no importers). Nothing in the library or the tests referenced any of them.
 
+### Deprecated
+
+- **`edgar.effect` and `edgar.form144` moved, with the old paths kept until 6.0.** `Effect` now lives in `edgar.offerings.effect`, beside the S-1/S-3/S-4, 424B and Form C/D parsers — a notice of effectiveness is a registration-lifecycle document. `Form144` now lives in `edgar.ownership.form144`, beside the Forms 3/4/5 parsers — Form 144 is an affiliate's notice of a proposed sale. Neither name was in `edgar.__all__` nor re-exported at the top level, so the module path was their whole public surface; both old paths still work and emit a `DeprecationWarning` naming the new one. The objects are the same objects, not copies, so `isinstance` checks against either path agree. `filing.obj()` routes to the new paths, so ordinary use never triggers the warning. (bead edgartools-07lk.12.1)
+
 ### Changed
 
 - **The wheel no longer ships development tooling.** `[tool.hatch.build].include` opens with an unrestricted `edgar/**/*.py`, so every install also downloaded the evaluation harnesses, entity training scripts and demos under `edgar/` — about 8,900 lines across `ai/evaluation/`, `ai/examples/`, `entity/training/` and `thirteenf/demo_comparison.py`. They are excluded now; `ai/exporters/` stays, since `export_skill` is public API. A regression test asserts that no shipped module imports any excluded path, which is the property that makes the exclusion safe rather than the exclusion itself.
